requirements.txt
autogen
pyautogen-agentchat
pyautogen-core
autogen-ext[docker]
python-dotenv
openai
tiktoken
docker
streamlit